Add definitions utilities.
