+++ safe/gui/tools/wizard/step_kw00_purpose.py
@@ -24,12 +24,13 @@ from PyQt4.QtCore import pyqtSignature
 from PyQt4.QtGui import QListWidgetItem,QPixmap
 
 from safe.definitionsv4.layer_purposes import layer_purpose_aggregation
+from safe.definitionsv4.utilities import purposes_for_layer
+
 from safe.gui.tools.wizard.wizard_step import WizardStep
 from safe.gui.tools.wizard.wizard_step import get_wizard_step_ui_class
 from safe.gui.tools.wizard.wizard_strings import category_question
 from safe.utilities.keyword_io import definition
 from safe.utilities.resources import resources_path
-from safe.gui.tools.wizard.wizard_utils import purposes_for_layer
 
 FORM_CLASS = get_wizard_step_ui_class(__file__)
