+++ safe/gui/tools/wizard/test/test_wizard_util.py
@@ -1,17 +1,13 @@
+# coding=utf-8
+
+"""Test for Wizard Util.
+"""
+
 import unittest
 
-from safe.gui.tools.wizard.wizard_utils import purposes_for_layer
 
 class TestWizardUtil(unittest.TestCase):
-    def test_layer_purpose_for_layer(self):
-        expected = ['aggregation', 'exposure', 'hazard']
-        self.assertListEqual(expected, purposes_for_layer('polygon'))
-
-        expected = ['exposure', 'hazard']
-        self.assertListEqual(expected, purposes_for_layer('raster'))
-
-        expected = ['exposure']
-        self.assertListEqual(expected, purposes_for_layer('line'))
+    pass
 
 
 if __name__ == '__main__':
